refactor(eval): Split out model configs to a separate file
